Extract Import/Export window to its own module and struct
